backend/requirements.txt
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
requests==2.31.0
